+++ src/Components/main/detailHeader/DetailHeader.styled.js
@@ -1,50 +1,4 @@
-// React import
-import React, { Fragment } from 'react';
-
-// style import
 import styled from 'styled-components';
-
-// Package import
-import { useNavigate } from 'react-router-dom';
-import { FaAirbnb } from 'react-icons/fa';
-import { FaBars } from 'react-icons/fa';
-
-
-// import ProfileModal from "./ProfileModal";
-
-const DetailHeader = ({ isScrollTop, togglePopup, isLoggedIn, data ,Filter,setFilter}) => {
-  // const { profileImg } = data;
-  const navigate = useNavigate();
-
-  return (
-    <Fragment>
-      <HeaderBox>
-        <HeaderLogo>
-          <FaAirbnb />
-          <HeaderLogoTitle onClick={() => navigate('/')}>
-            airbnb
-          </HeaderLogoTitle>
-        </HeaderLogo>
-        <HeaderPageButton
-          btnType="oval"
-          isScrollTop={isScrollTop}
-          // onClick={openProfileModal}
-        >
-          <FaBars/>
-          {/* {isLoggedIn && profileImg && <HeaderProfileImg src={profileImg} />}
-          {isLoggedIn && !profileImg && <HeaderProfileImg src={imgPath} />} */}
-          {!isLoggedIn && (
-            <HeaderProfileSvg>
-              <path d="m16 .7c-8.437 0-15.3 6.863-15.3 15.3s6.863 15.3 15.3 15.3 15.3-6.863 15.3-15.3-6.863-15.3-15.3-15.3zm0 28c-4.021 0-7.605-1.884-9.933-4.81a12.425 12.425 0 0 1 6.451-4.4 6.507 6.507 0 0 1 -3.018-5.49c0-3.584 2.916-6.5 6.5-6.5s6.5 2.916 6.5 6.5a6.513 6.513 0 0 1 -3.019 5.491 12.42 12.42 0 0 1 6.452 4.4c-2.328 2.925-5.912 4.809-9.933 4.809z"></path>
-            </HeaderProfileSvg>
-          )}
-        </HeaderPageButton>
-      </HeaderBox>
-    </Fragment>
-  );
-};
-
-export default DetailHeader;
 
 export const HeaderBox = styled.header`
   max-width: 1120px;
@@ -61,7 +15,6 @@ export const HeaderBox = styled.header`
   margin-left: auto;
   margin-right: auto;
   margin-top: 10px;
-
 `;
 
 export const HeaderLogo = styled.div`
