+++ src/Pages/Main/Main.js
@@ -3,13 +3,12 @@ import React, { Fragment, useState, useEffect } from 'react';
 import { useSelector } from 'react-redux/es/exports';
 import { useDispatch } from 'react-redux/es/exports';
 import { MainThunk } from '../../redux/Modules/PageModules/Main';
-import { CategoryThunk } from '../../redux/Modules/PageModules/Category';
 import { useNavigate } from 'react-router-dom';
 
 // components import
-import Header from '../../Components/main/Header';
-import Card from '../../Components/main/Card';
-import Category from '../../Components/main/Category';
+import Header from '../../Components/main/header/Header';
+import Card from '../../Components/main/card/Card';
+import Category from '../../Components/main/category/Category';
 import MainSkeleton from '../../Components/skeleton/MainSkeleton';
 import FilterModal from '../../Components/Modals/FilterMordal/FilterMordal';
 
@@ -18,12 +17,10 @@ import {
   FilterButton,
   CategoryBox,
   CategoryGroup,
-} from '../../Components/main/Category';
+} from '../../Components/main/category/Category';
 
 // style import
 import styled from 'styled-components';
-import { useParams } from 'react-router-dom';
-// import ImageSlider from '../../Components/main/ImageSlider';
 
 const Main = () => {
   const dispatch = useDispatch();
